Version 2
In this version, i have added the labels for the grid, asked player 1 to input the row and column they want to place their disk in, checked to see if the location was in the grid and if that location was empty. If both aren't true, then the player is asked to try again.